requirements.txt
openai>=1.30.0
playwright>=1.44.0
fpdf2>=2.7.0
python-docx>=1.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
plyer>=2.1.0
playwright-stealth>=1.0.6
